Test on click event
