+++ about/build.gradle
@@ -32,27 +32,29 @@ android {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
+
 }
 
 repositories {
     google()
+    mavenCentral()
     jcenter()
 }
 
 dependencies {
     implementation project(':app')
-    implementation "android.arch.lifecycle:viewmodel:${versions.lifecycle}"
-    implementation "android.arch.lifecycle:extensions:${versions.lifecycle}"
-    implementation "com.android.support:support-fragment:${versions.supportLibrary}"
 
     testImplementation project(':test_shared')
+    testImplementation "android.arch.core:core-testing:${versions.lifecycle}"
+    testImplementation "com.nhaarman:mockito-kotlin:${versions.mockito_kotlin}"
     testImplementation "junit:junit:${versions.junit}"
-    testImplementation "org.mockito:mockito-core:${versions.mockito}"
 
+    androidTestImplementation project(':test_shared')
+    androidTestImplementation "android.arch.core:core-testing:${versions.lifecycle}"
     androidTestImplementation "com.android.support.test:runner:${versions.test_runner}"
     androidTestImplementation "com.android.support.test:rules:${versions.test_rules}"
     androidTestImplementation "com.android.support.test.espresso:espresso-core:${versions.espresso}"
-    androidTestImplementation "org.mockito:mockito-core:${versions.mockito}"
+    androidTestImplementation "com.nhaarman:mockito-kotlin:${versions.mockito_kotlin}"
     androidTestImplementation "org.mockito:mockito-android:${versions.mockito}"
-    androidTestImplementation "com.squareup.retrofit2:retrofit-mock:${versions.retrofit}"
+    androidTestImplementation "org.mockito:mockito-core:${versions.mockito}"
 }
